Split LG and LGNode classes
